# All Voucher Amounts should show Dr. / Cr. -> Done # Card dimensions changing on search -> Done # User should be able to enter amount in decimals -> Done # Layover try "Overnight layover in Thiruvananthapuram" - put in 2 lines  -> Done # Style and spacing absolutely different from design -> Done # website missing  -> Done # Hotel phone number with country code and flag should appear -> Done # Add to apple wallet functionality missing. Loading style also incorrect - should be loader on button like social icons -> Done # Email chip: -> Done # Room filters should have 16px right margin with dot -> Done

+++ AERTRIP/Models/BookingDetail/BookingFeeDetail.swift
@@ -11,21 +11,21 @@ import Foundation
 struct BookingFeeDetail {
     
     struct Charges {
-        var adult: Int?
-        var child: Int?
-        var infant: Int?
+        var adult: Double?
+        var child: Double?
+        var infant: Double?
         
         init(json: JSONDictionary) {
             if let obj = json["ADT"] {
-                self.adult = "\(obj)".toInt
+                self.adult = "\(obj)".toDouble
             }
             
             if let obj = json["CHD"] {
-                self.child = "\(obj)".toInt
+                self.child = "\(obj)".toDouble
             }
             
             if let obj = json["INF"] {
-                self.infant = "\(obj)".toInt
+                self.infant = "\(obj)".toDouble
             }
         }
     }
@@ -100,7 +100,7 @@ struct AerlineCharge {
 struct FEE {
     var from: Date?
     var to: Date?
-    var value: Int?
+    var value: Double?
     
     init(json: JSONDictionary) {
         if let obj = json["from"] {
@@ -112,7 +112,7 @@ struct FEE {
         }
         
         if let obj = json["value"] {
-            self.value = "\(obj)".toInt
+            self.value = "\(obj)".toDouble
         }
     }
